A simple website built using only HTML and CSS, linking free IT courses from YouTube via iframes for practice and learning.
This website is a basic project designed for practice using only HTML and CSS. It features a clean and straightforward layout where free IT courses from YouTube are embedded via iframes. The goal of this project is to provide easy access to a variety of learning resources while practicing web development skills. Perfect for those looking to enhance their web development abilities with hands-on experience in creating a functional, content-driven site.